requirements.txt
clickhouse-connect

pandas

langchain-community
langchain-core
langchain-openai

python-dotenv
faker
PyYAML
